+++ node/index.js
@@ -1,124 +1,119 @@
 var fs = require("fs");
 var util = require("util");
 
+// Common dependencies to get LayoutManager running.
 var Backbone = require("backbone");
 var _ = require("underscore");
-var def = require("underscore.deferred");
+
+// Using Cheerio instead of jQuery, because Cheerio emulates a loosely
+// compatible API that we further augment to ensure unit tests pass.  It is
+// also much faster than jsdom.
 var $ = require("cheerio");
 
-// Hacky way to get layoutmanager to load with the correct globals
-(function() {
+// This is to avoid unwanted errors thrown when using
+// `Backbone.View#setElement`.
+$.prototype.unbind = function() {};
 
-  eval(fs.readFileSync(__dirname +
-    "/../backbone.layoutmanager.js").toString());
+// Since jQuery is not being used and LayoutManager depends on a Promise
+// implementation close to jQuery, we use `underscore.deferred` here which
+// matches jQuery's Deferred API exactly.
+var def = require("underscore.deferred");
 
-  // Helper function to get a value from a Backbone object as a property
-  // or as a function.
-  var getValue = function(object, prop) {
-    if (!(object && object[prop])) {
-      return null;
-    }
+// Get Backbone and _ into the global scope.
+_.defaults(global, { Backbone: Backbone, _: _ });
 
-    return _.isFunction(object[prop]) ? object[prop]() : object[prop];
-  };
+// Include the LayoutManager source, without eval.
+require("../backbone.layoutmanager");
 
-  // Ensure Backbone doesn't ensureElement
-  Backbone.View.prototype._ensureElement = function() {
-    var attrs;
+// Set the Backbone DOM library to be Cheerio.
+Backbone.$ = $;
 
-    if (!this.el) {
-      attrs = getValue(this, "attributes") || {};
+// Need to override `make` here, because Backbone does not create the element
+// with the DOM library.
+//
+// Marked for removal, once new version of Backbone is released that removes this function.
+Backbone.View.prototype.make = function(tagName, attributes, content) {
+  var $el = Backbone.$("<" + tagName + ">");
+  if (attributes) $el.attr(attributes);
+  if (content != null) $el.html(content);
+  return $el[0];
+};
 
-      if (this.id) {
-        attrs.id = this.id;
+// Configure LayoutManager with some very useful defaults for Node.js
+// environments.  This allows the end user to simply consume instead of
+// fighting with the desirable configuration.
+Backbone.Layout.configure({
+  deferred: function() {
+    return def.Deferred();
+  },
+
+  fetch: function(path) {
+    var done = this.async();
+
+    fs.readFile(path, function(err, contents) {
+      if (err) {
+        console.error("Unable to load file " + path + " : " + err);
+
+        return done(null);
       }
 
-      if (this.className) {
-        attrs["class"] = this.className;
-      }
+      done(_.template(contents.toString()));
+    });
+  },
 
-      this.setElement(this.make(this.tagName, attrs), false);
-    } else {
-      this.setElement(this.el, false);
-    }
-  };
+  // This is really the only way you will want to partially apply a view into
+  // a layout.  Its entirely possible you'll want to do it differently, so
+  // this method is available to change.
+  partial: function(root, name, el, insert) {
+    // If no selector is specified, assume the parent should be added to.
+    var $root = name ? $(root).find(name) : $(root);
 
-  Backbone.View.prototype.make = function(tagName, attributes, content) {
-    var el = $("<" + tagName + "/>");
-    if (attributes) el.attr(attributes);
-    if (content) el.html(content);
-    return el;
-  };
-
-
-  Backbone.View.prototype.setElement = function(element, delegate) {
-    this.$el = element;
-    this.el = element[0];
-
-    return this;
-  };
-
-  Backbone.LayoutManager.configure({
-
-    manage: true,
-
-    deferred: function() {
-      return def.Deferred();
-    },
-
-    fetch: function(path) {
-      var done = this.async();
-
-      fs.readFile(path, function(err, contents) {
-        if (err) {
-          console.error("Unable to load file " + path + " : " + err);
-          return done(null);
-        }
-
-        done(_.template(contents.toString()));
-      });
-    },
-
-    // This is really the only way you will want to partially apply a view into
-    // a layout.  Its entirely possible you'll want to do it differently, so
-    // this method is available to change.
-    partial: function(root, name, el, append) {
-      // If no selector is specified, assume the parent should be added to.
-      var $root = name ? $(root).find(name) : $(root);
-
-      // If no root found, return false.
-      if (!$root.length) {
-        return false;
-      }
-
-      // Use the append method if append argument is true.
-      this[append ? "append" : "html"]($root, el);
-
-      // If successfully added, return true.
-      return true;
-    },
-
-    // Override this with a custom HTML method, passed a root element and an
-    // element to replace the innerHTML with.
-    html: function($root, el) {
-      $root.html(_.isString(el) ? el : $(el).html());
-    },
-
-    // Very similar to HTML except this one will appendChild.
-    append: function($root, el) {
-      $root.append(_.isString(el) ? el : $(el).html());
-    },
-
-    when: function(promises) {
-      return def.when.apply(null, promises);
-    },
-
-    contains: function() {
+    // If no root found, return false.
+    if (!$root.length) {
       return false;
     }
 
-  });
+    // Wrap the supplied element as a Cheerio object to ensure that insertion
+    // does not trigger the creation of a new node. This maintains object
+    // constancy across "DOM" operations so that the "contains" method can
+    // recognize identical nodes.
+    el = $(el);
 
-}).call(_.extend(global, { Backbone: Backbone, _: _ }));
+    // Use the insert method if `insert` argument is true.
+    if (insert) {
+      this.insert($root, el);
+    } else {
+      this.html($root, el);
+    }
 
-module.exports = Backbone.LayoutManager;
+    // If successfully added, return true.
+    return true;
+  },
+
+  when: function(promises) {
+    return def.when.apply(null, promises);
+  },
+
+  contains: function(parent, child) {
+    var $child = $(child);
+
+    // According to the jQuery API, an element does not "contain" itself.
+    if (child === parent) {
+      return false;
+    }
+
+    // Step up the descendents, stopping when the root element is reached
+    // (signaled by `.parent()` returning a reference to the same object).
+    while ($child !== $child.parent()) {
+      $child = $child.parent();
+
+      if ($child[0] === parent) {
+        return true;
+      }
+    }
+
+    return false;
+  }
+});
+
+module.exports = Backbone.Layout;